Generate plots
+ Rejection sampling takes as example a "complicated" function f(x) created
by the product of a Gaussian with a absolute sine wave. It is subsequently
displaying the histogram of the distribution generated by rejection
sampling and plotting both f(x) and the simple distribution g(x) which
is in this case the original Gaussian (we know it is equal or larger
than f(x) because the latter is a product with a sine function <= 1).
+ The pareto pair distribution can generate points on a line between two
endpoints. We show the endpoints are adjusted the more samples we have.
Hence, not just sampling, but calculating the posterior for the
endpoints with more and more samples.
+ Segments generated from a Dirichlet process using a Normal Inverse
Wishart combined with a Pareto distribution as base distribution and using a
Polya urn model to decide if we have to sample on old or a new base
distribution (with alpha = 1).